+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - ``mpserver.Worker`` got new parameter ``worker_index`` to ``__init__``, which is
   automatically provided by the parent ``ProcessServlet`` or ``ThreadServlet``.
+  Subclasses of ``Worker`` should be sure to accept this parameter in their ``__init__``.
 
 
 ## [0.12.3] - 2023-04-14